[TASK] styleguide: Add commands for TCA and frontend create/delete
Add the following command:
./bin/typo3 styleguide:generate frontend|tca|all

With the option '--delete' and '--create' to manage
page trees and records.

Move 'styleguide:kauderwelsch' command to Services.yml

+++ typo3/sysext/styleguide/Classes/TcaDataGenerator/Generator.php
@@ -199,8 +199,10 @@ class Generator extends AbstractGenerator
         $this->deleteFalFolder('styleguide');
 
         // Delete site configuration
-        $site = GeneralUtility::makeInstance(SiteFinder::class)->getSiteByRootPageId($topUids[0]);
-        GeneralUtility::makeInstance(SiteConfiguration::class)->delete($site->getIdentifier());
+        if($topUids[0] ?? false) {
+            $site = GeneralUtility::makeInstance(SiteFinder::class)->getSiteByRootPageId($topUids[0]);
+            GeneralUtility::makeInstance(SiteConfiguration::class)->delete($site->getIdentifier());
+        }
     }
 
     /**
